BGE: fix animation waking up for suspended scenes.
rBSd979ac4cc9967c5dce27461926396abe5de54b3d introduced artefact with animations on suspended scenes.
The problem come form the suspended delta time (scene pause time) set in the main loop.
In the case of 2 scenes, if the first is suspended, not proceed and that the second resume it, the pause time will be not set for the first scene. The cause of wrong pause time create in BL_Action::Update the artifcats, indeed the action do `localtime = curtime - pausetime` to manage properly pauses but in this case pausetime = 0 and for the action there was no scene's pause. It's like the action was played during the scene's pause.
To fix it the patch just set pause time properly by explode the main loop, this allow to get the proper scene status (suspended/resumed) after the main loop and set the pause time cooresponding.

Example file : {F246049} : press S to suspend and R to resume.

Reviewers: moguri, lordloki, youle, sybren

Subscribers: dfelinto

Differential Revision: https://developer.blender.org/D1569