UPBGE: Fix scene deletion in converter RemoveScene function.
The converter need to free the scene before free all data like meshes and materials.
To solve this each time we call KX_BlenderSceneConverter::RemoveScene we remove the scene form the engine scene list without release it.
To egalize the reference count to 1 (only the scene list is the owner) we must release the startscene in main loop.

Thanks to youle for the bug repport.