UPBGE: Fix wrong viewport during PostRenderScene.
Previously debug draw like physics was drawn incorrectly because of a wrong viewport in blenderplayer, stereo and multiple viewport mode.
In fact the viewport was set to the fully render area for post scene render, so every post scene render get the same viewport. But this
behaviour is useful for 2d filter. To fix it the scene viewport is keeped before rendering filter and set again after.

This patch was tested in both blenderplayer and blender with different modes:
- physics debug draw
- viewport
- 2d filter
- stereo (each stereo mode)
- oculus 2d filter