UPBGE: Fix render with custom shader.
Previously we rendered only mesh with a multitexture or blender glsl material,
but a custom glsl shader is not one of it, so it was rendered like single texture material.
Since single texture render function was removed, custom shader was not rendered.
To fix it the condition was removed to render all meshes which are not text.