BGE: Add bounding volume support for the user, Fix T33191.
Previously when a mesh was deformed the bounding volume or AABB was not updated and it caused some wrong frustum culling.
To fix it i added an AABB compute in RAS_MeshObject and in RAS_Deformer, which are updated each time the mesh or deformer is modified but as Moguri noticed it can spent 10% spend more time than deforme the mesh.
So the solution is to add an option to automatically update or not the AABB or use a predefined AABB based on a mesh.
It fixes T33191.
{F226655}

Reviewers: moguri, campbellbarton, hg1, agoose77, lordloki, Matpi, youle

Subscribers: blueprintrandom, dfelinto

Maniphest Tasks: T33191

Differential Revision: https://developer.blender.org/D1475