UPBGE: Fix T41677: Derived mesh not rendered with no face original mesh.
Previously the only way to add a material in a mesh was to add polygon in it, but for meshes without polygons and with a modifier which creates polygons,
none material was added. To fix it the first material is converted if there's no polygons by using a new function : RAS_MeshObject::AddMaterial.
Checks around tface, mface was added in functions about material conversion.