UPBGE: Fix T24948: Material not fully converted with modifiers.
Previously we used RAS_IPolyMaterial::GetMaterialIndex, but this index is wrong, because if there's two materials and two objects: the first contains
the blue material and red material and the second the red material.
During the consversion of the first object, the blue material's index is set to 0 and the red to 1, but in the second object conversion, the red
material's index is set to 0.
It will cause that during render the red and blue material were considerated with the same index and it break DerivedMesh render which use material
index.

Notes:
- A hacky fix for it was removed in RAS_MeshObject::AddMeshUser
- all use of RAS_IPolyMaterial::GetMaterialIndex must be removed.