UPBGE: Debug draw armature python function.
A new function in the python API BL_ArmatureObject.draw() was implemented to debug draw armature's bones with simple red lines.
In the C++ side the function BL_ArmatureObject::DrawDebugArmature is added, and a boolean BL_ArmatureObject::m_drawDebug.
The major issue is that armatures are updated after the python step. To fix it the python function set m_drawDebug to true and
in DrawDebugArmature the render is does only if m_drawDebug is true and in this case set to false at the end of the function.

This new mode allow the user to see what constraint does, but for the developer is a good way to see bugs.