UPBGE: NULL vertex array update for modifier deformer.
Modifier deformer try to update it's vertex array via function UpdateTransverts, but modifier deformer doesn't have one.
Previously vertex array was not set to NULL for modifier deforme and it was the one from the orignal mesh. It means that even if it's
useless modifier did a copy to a vertex array which can be freed !
To solve we abort function if UseVertexArray() return false.