UPBGE: Fix zombie object for removed child during conversion.
If a child object is not in the same layer than the parent it will be removed, but due to recent modification about temporary list returned by
GetChildren and GetChildrenRecursive, these list are temporary so not owned and it's item should be the same.
To solve the issue a counter increase AddRef was removed.