UPBGE: Optimize bounding box radius.
Previously the bounding box radius (used to optimize non Bullet culling process) was based on the object center, which cause for uncentered
mesh an huge box radius. To solve this issue 2 new functions was implemented in : SG_BBox::GetCenter() SG_BBox::GetRadius(), these function
return the center of the box (max - min) / 2 and the radius (max - min).length(). So now the culling sphere is centered on the AABB and use
the proper AABB radius.