UPBGE: Fix scene replacement and deletion.
After call m_sceneconverter->RemoveScene(scene) nobone have to acces to the scene, the scene must be freed in the called function.
If the scene is created and inserted  with list->InsertFunc(scene->AddRef()) in the same function, it should be release because of
the original reference coutn set to one at constructor level.