UPBGE: optimize display list recompile for deformer.
The deformer can return true when Aplly is called, this value mean that the deformer data was copied in the mesh slot's vertex array.
The wrong behviour is to say that the mesh object is modified while is only on of its derivated vertex array.
It not really cause a bug but theorically a speed down, if you have a 2 objects with the same mesh and that one of it uses a deformer
like soft body, the other one normal can't use any optimization because is mesh is considerate as modified every times.