UPBGE: Optimize auto bound update, remove auto update UI option.
The auto update value is now set to true if there's no predefined bound mesh.
Previously if auto update was to true, we always get and set the AABB, now we do that only if the there's auto bound update and the mesh is modified
or the mesh deformer is dynamic.
The only bad side of this is that the AABB is not update when changing of mesh, to solve this UpdateBounds now take a param to force the update.