UPBGE: Fix T37125, modified mesh not updated to deformed objects.
Previously there was no copy of original mesh data like UVs or colors to the unique deformed mesh for armatures softbody and shapekeys.
To solve it, BL_SkinDeformer::Apply now copy UVss, colors and tangent if the mesh is modified.
But not all info should be copied if the meshis modified, to optimize it a new system was added.

Now RAS_MeshObject separate its modification by category stored in a flag set by SetModifiedFlag, altereted by AppendModifiedFlag and get
by GetModifiedFlag.
By doing this the BL_SkinDeformer can know which mesh data are modified and should be updated in the unique deformed mesh.