UPBGE: Fix armature not drawed when no visible children.
Previously the debug draw was done in the ApllyPose function which is called only when the armature is updated, but the armature is updated only if it has
children to optimize.
To solve this we must move the debug draw function outside ApplyPose in a new function KX_Scene::DrawDebug which is merged with the KX_Scene::DrawBoundingBox
function.
To make the armature debug draw like the boudnign box debug draw, the "do nothing" check was moved outside of BL_Armature::DrawDebugArmature() by using
BL_ArmatureObject::GetDrawDebug().