UPBGE: Fix multithread conflict on RAS_TexVert::xyz when computing deformer AABB.
RAS_TexVert::xyz use a workaround to optimize by keeping reference: it updates its position on a static MT_Point3 (g_pt3) and return a reference on it.
This behavior disallow two thread to call xyz() at the same time, even more if the returned value is stored as a reference by the caller.
To solve this the function now return a fully new MT_Point3 based on the copied value of the position, it also make the code more understandable.