UPBGE: Fix unshared messages between scenes.
In the message/network refactor this behaviour was broken, to restor it all KX_NetworkMessageScene send their message to a KX_NetworkMessageManager owned by the ketsji engine.
All messages in this manager are removed at the end of a full logic frame.
By doing this KX_NetworkMessageScene seems useless, currently it only does an interface between the sender and the message manager. This interface will be
useful later when dealyed message will be implemented, indeed it will disallow sending delayed message if the scene is suspended, and in this case the
KX_NetworkMessageScene will store this kind of messages.