UPBGE: Remove material pass and fake call to ActivateMaterial.
The old pass sytem obligate to bind the material after binding vertex array attribut, but the function
which activate the material disable the last one and them disable textures with it, by disabling
textures all vertex array textures coordinate was broke.

To fix it the pass system was removed, this system was unused by blender material and doens't seems useful
for anything since the BL_Shader.setNumberOfPasses(max_pass) documentation said that the function was not
implemented.

The use of global last shader was dangerous too, a safe global variable doesn't change of value. To replace
it the code use now explicit call to ActivateMaterial and DesatcivateMaterial, by doing this we make sure
the material is activated and desactivated after a render. The side effect of this is to simplifacte a lot
all KX_BlenderMaterial functions, indeed most of theses functions checked if there's a different last shader
and desactivate it.

Because of the remove of fake active material call, the function which set all lighting attribut was moved
to RenderMeshSlot.

Despite these changes the render is now slowed of 2.6 % for 1000 meshes of 4 materials.