UPBGE: Use RAS_IPolyMaterial::UsesObjectColor instead of RAS_MeshSlot::m_bObjectColor.
Previously we mixed this function and this variable which can cause some not understandable. To make things clearer UsesObjectColor() return
true if the material use the object color option and if the material is a not a blender glsl one. Indeed the returned value is used to bind
the object color via glColor for customs shaders and multitexture.