UPBGE: Optimize quadrilateral to triangles conversion.
Previously this conversion treated all vertex as duplicate, but during the conversion we already know which vertex are shader by the two
triangles resulting of a quadrilateral.

To avoid check on these shared vertex we changed a bit the way of adding polygones and vertexes. Now RAS_MeshObject::AddVertex take the material
bucket instead of the polgygon and return the index of the added or existing vertex. RAS_MeshObject::AddPolygon changed too, now it takes an
array of 4 vertex indices and visibile, collider, twoside flags, this function does the quadrilateral to triangles conversion internally.