UPBGE: Fix whitescreen when undefined variable in 2D filter
In Blender official, when there is an error in 2D filter shader, the
filter is not executed. In upbge, the shader was nevertheless executed
causing a white screen. This patch disables the execution of the shader if
an error is detected. The message error is thrown, but the filter is not
executed and you can see the scene without the filter applied.