UPBGE: Implement parallax mapping.
Parallax mapping is fake dynamic texture which gives us the impression of complexe models. Useful for wall, rock, ground.

This patch is a simplified version of Angelo Miner patch about parallax mapping. If this option is enabled in texture panel
the user have to pass a texture with a alpha channel cooresponding to depth of the texture. The UI for parallax is
composed of 5 values :
- Use Parallax UV
- Height Scale
- Step
- Bump Scale
- Discard Edges

Authors: Angelo Miner, Ulysse Martin, Diego Lopes.