UPBGE: Setup conversion system from older/non UPBGE file version.
Previously there was no difference between a UPBGE file version and a regular one, it cause some double conversion on file
that doesn't requierd it.
The UPBGE version is now treated as an independant file version system, to do this upbge[sub]version[file] was added in
Main struct and FileGlobal struct. The new conversion function is named blo_do_versions_upbge contained in versioning_upbge.c.
This function do all kind of conversion to UPBGE file. If the file was written with regular blender, the version and subversion
will be to 0, else it will be like 0 and 3 for the moment.
Two macro was added for the UPBGE version/subversion : UPBGE_[SUB]VERSION to 0 and 3.