UPBGE: Allow get render mesh transform matrix without apply it to OpenGL
Previously RAS_IRasterizer::applyTransform compute the matrix and applied it directly.
For the regular render draw it's fine, but for the soon geometry instancing render it's bad
because it obligate us to use glGetFloat and disallow compute matrix in parallale tasks.

To solve that we separate the old RAS_IRasterizer::applyTransform in two functions :
RAS_IRasterizer::GetTransform(float *origmat, int objectdrawmode, float mat[16])
RAS_IRasterizer::ApplyTransform(const float mat[16])

GetTransform only compute the matrix for the given drawing mode and copy the result to "mat" argument.
ApplyTransform only call glMultMatrixf with the given matrix.

The side effect of compute the matrix independently is that the shadow draw mode had to get the orignal
matrix and final matrix in the ray cast function info argument. To do that th new struct RayCastTransform
was declarated, it contains the original matrix "origmat", the output matrix "mat" and the object size
"scale". An instance of it is passed in KX_RayCast::Callback<RAS_OpenGLRasterizer, RayCastTranform>.

The original code was a bit cleaned especially for the translation and "size" to "scale".

Due to the separation the matrix computation is a bit slower (2%) for normal meshes and faster for billboard meshes,
but it will allow use later to use optimizations like geometry instancing.