Move instancing buffer managment in RAS_InstancingBuffer and support transparency sort.
All function around the instancting buffer was moved in RAS_InstancingBuffer. This class contains
two function to bind/unbind, a function to allocate and fill the VBO and 4 functions to get the
stride, position, matrix and color pointer.

An instance of this class is created in RAS_DisplayArrayBucket::RenderMeshSlotsInstancing if
it not already exists, and deleted in the RAS_DisplayArrayBucket destructor.

The mesh slot sort is based on the RAS_BucketManager sort code, but to reuse it we had to move
the struct declaration in the header file, for struct like sortedmeshslot, backtofront,
fronttoback.

In RAS_DisplayArrayBucket::RenderMeshSlotsInstancing a new argument "alpha" is passed, if this
argument is to true we store the mesh slot in sortedmeshslot instance, sort them and store the
result in a RAS_MeshSlot list, after we pass to RAS_InstancingBuffer::Update() the sorted mesh
list. Else if "alpha" is to false, we passed the normal mesh slot list : m_activeMeshSlot.