UPBGE: Support sky/background render.
Previously the sky render was not supported, this render can be useful with cubemap to replace a skybox.

To do this we first expose world UI option for the BGE.
In the rasterizer a new function RAS_OpenGLRasterizer::RenderBackground() was added to draw a quad with a
always true depth test.
In the KX_WorldInfo class the blender scene is now stored and the new function
KX_WorldInfo::RenderBackground(RAS_IRasterizer *) bind the gpu world material and call the backgound render
to the rasterizer.

Only the first scene do the background render for its world to avoids issue and add an option.