UPBGE: Fix wire material.
Previously all wire material was using triangles and not quad as BF BGE. The perfect behavior
should be to respect edges and not only faces.

Now we use only lines not quad or triangles and these lines are real edge of valid faces.
To find a edge we test if two vertices are en edge in the poly which contains the face,
if yes we call RAS_MeshObject::AddLine(bucket, v1, v2). And in AddPolygon the vertex
offset are don't added if the material is wire : RAS_IPolyMayerial::IsWire().

Now in RAS_DisplayArray we have a new variable m_type depending of an enum which contains
LINES, TRIANGLES. The function GetOpenGLPrimitiveType convert this value to an opengl
primitive type.

RAS_IRasterizer::SetLines became useless for regular meshes (because they use GL_LINES
directly) but it's still resquierd for DerivedMesh. The call to this function was removed
in KX_BlenderMaterial and added in RAS_OpenGLRasterizer::DrawDerivedMesh.