UPBGE: Fiw zombie object for parented group object in hidden layer.
All group object are added in BL_BlenderDataConverion to a list named convertedlist, but
they are not removed from this list when the object is in hidden layer and a children.
To solve it we first convert convertedlist to a CListValue for the useful functions
FindValue, RemoveValue ect... After we add a check and remove in hidden children deletion
code part.