UPBGE: Fix color in wireframe render.
Previously the vertex color pointer was alwyas set in VBO and VA mode and the dummy instancing
fragment shader was useing a white color.
To fix this we changed the color in the shader to vec4(0, 0, 0, 1).
About VA and VBO we add condition to set the vertex color pointer, if it's in wireframe it does
nothing. For the VBO is reqested a new argument in function Bind and Unbind to now the drawing
mode.