UPBGE: Implement PCF shadow and sharpness variance shadow.
PCF is a shadow mapping technique used to create smooth shadow
edges by sampling multiple time in a small zone and mix the results.

Variance sharpness is a value representing the distance between samples
when blurring the color texture of the variance shadow.

PCF depends as for simple shadow of a depth texture with coparaison
enabled. The only difference are in the UI options showed and the shader
functions used.
For the UI the user can interact with the option "samples" for the root
squared number of sampler and the option soft for the distance between
samples. For the shader two functions are used for PCF: test_shadowbuf_pcf
and test_shadowbuf_pcf_early_bail.

The first do every samples and divide the summary. The second
functions test first if the fourth furthest samples, if the values are
the same (all 0.0 or 1.0), then we discard all other samples and use these
four values. It economize a lot of process but can create artifacts for
closest shadow edges.

As PCF introduced two new long functions, "shadows_only" is simplified to
reuse a shadow value computed from an other function described before
instead of calling itself these functions.

Variance sharpness is showed in the UI under the name of "Sharpness".
if it's value is close to -1 then the shadow edges are smooth else they are
hard.