UPBGE: Disable SSS for directional lights and reenable shadows on SSS material
This patch is a temporary solution to fix issues with SSS. SSS shader is
written only for point lights. The influence of directional lights or
spotlights is not computed correctly. It's a bit weird to have a material
in the scene that is not influenced by all lights, but as the influence of
these lights is incorrect, we disable it for the moment. So SSS must be used with a point
light.

But we reenable shadows received by SSS materials.