UPBGE: Implement python component.
Python components are used to replace some complex logic brick that can be performed easily in python.

As they are to replace logic brick they had a UI with property. a property can be:
- float
- int
- string
- vector[2/3/4]
- set

To define these property and the component we derived in python a class from the new type
bge.types.KX_PythonComponent.
This type contains two major functions : the function "start" and "update", as exepted "start"
is called to initialize the component and "update" to perform a action every logic frame.

The property exposed to the UI should be declared in the component class as dictionary or an
ordered dictionary :
arg = {
	"Property 1" : 0.0,
	"Property 2" : "objectname"
}

Here two properties will be drawed in the UI for a float and a string.

The script containing the component can be a text in the blend file or an extern module,
anyway the both will be runned outside the BGE to parse the properties. But we don't by
the BGE module during this parse because of difficulties to import the BGE (and C++) in
the blender kernel part which manage components.

To solve this we simply bind a fake module, submodule and class for
bge.types.KX_PythonComponent.
This also allow the user to make multiple subclasses.

To detect that we're running the component module or srcipt outside the bge, the "bge"
module contains a "__component__" attribut only outside the BGE.

Finally python component are exposed in KX_GameObject.components.