UPBGE: New lod system.
Previously all converted meshes used for lod was stored in each object. By doing this
we used a lot of memory and made a complex system for switching of lod list.

To solve this all the lod system was moved in a class named: KX_LodList.
This class contains a list of KX_LodList::Level which are mesh levels set by the user.

So know in KX_GameObject::UpdateLod we simply use KX_LodList::GetLevel which return the
cooresponding level to distance and previous level index.

One of the advantage of KX_LodList is that it will possible to switch of lod an object.
For example in python:
obj1.lodList = obj2.lodList

Used in complement of a replaceMesh.