UPBGE: Implement thread lock classes.
The interest of thread lock class is to create a global instance for a lock
without care about it's initialization and deletion at the program end.

To do it we created two types of lock : Mutex and SpinLock, which represent
CThreadLock and CThreadSpinLock the both are a subclass of the pure virtual
class CThreadLock. This class contains the two major functions:
- CThreadLock::Lock()
- CThreadLock::Unlock()