UPBGE: Fix soft body shared vertex arrays.
Since we apply deformer before render to avoid mixing code stage, the soft body system doesn't
work, because this system was reusing always the same vertex array (the original one).
To fix it, we now disallow sharing and use the soft body index form the original vertex array,
because soft body vertex index are written only in the original vertexes.