UPBGE: Implement KX_WorldInfo.[zenithColor/horizonColor].
Previously the background color was broken since the sky render support
because the material overdraw the background color.

To fix it we now support colors modifications for the world info, but the
world doesn't take only one background color but two colors for the horizon
and zenith, theirs python API implementation is:
scene.world.zenithColor
scene.world.horizonColor

From ImageRender:
imagerender.zenith
imagerender.horizon

Due to these new values, the KX_WorldInfo.backgroundColor is now removed
and the UI option to use it (use background render) is also removed.

Also the color variables are now normalized and not between 0.0 and 255.0.

Since the background color is removed and not used in glClearColor, we now
pass to glClearColor 0.0s.