UPBGE: Refactor rasterizer opengl call.
The goal of this commit is to reduce opengl call outside the Rasterizer/
code part.

To do this we have to replace each opengl call by a function accesible from
RAS_IRasterizer.

The replaced functions are:
- gl[Enable/Disable] -> [Enable/Disable](EnableBit)
- glDepthFunc -> SetDepthFunc(DepthFunc)
- glBlendFunc -> SetBlendFunc(BlendFunc, BlendFunc)
- glClear -> Clear((int)ClearBit)
- glViewport -> SetViewport(int, int, int, int)
- glScissor -> SetScissor(int, int, int, int)
- glGetIntegerv(GL_VIEWPORT...) -> GetViewport(int *)

Also some internal function are used only in RAS_OpenGLRasterizer like:
- LoadMatrix(float *)
- LoadIdentity
- SetMatrixMode(MatrixMode)

To summarize every instances which want to call opengl function must
contains a rasterizer pointer. Which was not the case for RAS_ICanvas
and subclasses, to solve this now RAS_ICanvas constructor take as argument
a pointer of the rasterizer.

All RAS_IRasterizer function exposed before use enum, but these enum are
not the OpenGL value bcause it will, in this case, requierd to include glew-mx.h.
So the issue here is to convert the enum value to the OpenGL enum value.
To do this we use 4 int array table:
- openGLEnableBitEnums
- openGLDepthFuncEnums
- openGLMatrixModeEnums
- openGLBlendFuncEnums

And then get the OpenGL value by get the enum value index item, for example:
openGLEnableBitEnums[RAS_TEXTURE_2D] -> GL_TEXTURE_2D

By doing this we allow a fast convertion, but any new added enum should respect the same
order set in the enum declaration in RAS_IRasterizer and the int array in
RAS_OpenGLRasterizer.