Implement properties reuse for components reload.
The component load form python was moved to "load_component", this function is called after
create the component in new_component_from_module_name or directly in reload_component.

With this move the property reuse was added, for each property we checked if we don't find a
coresponding one in the old properties list, if yes we removed this old property from its list
and add it in the new properties list, the new created property is released. Else we simply
add the new property.
At the end we free all properties which are rest in the old properties list.