Fix uninitialized properties list.
Previously the properties list was using fake poniter for first and
last item instead of NULL, this was causing a segfault because adding
an item need to add a pointer of it in the previous and next items.

To fix it we use memset to set all values in the ListBase to 0.