UPBGE: Allow KX_GameObject rename.
Previoulsy it was impossible to rename a object even a replicated object.
Now KX_GameObject.name is available in write mode.
A replicated object is renamed without any issues but a non-replicated object
should be unregister for its old name and re-register with the new name.
This could not be possible it an other non-replicated object has the same
name, in this case we raise an error.

The unregister and register call are here to keep use of functions like
scene.addObject("NewName")