UPBGE: Refactor BL_Texture.
Previously texture conversion for custom shaders was done manually in BL_Texture
and it was binding the texture manually.

To simplify the code we reuse GPU_texture functions to create and manage texture
bind. So now the BL_Texture stores the GPUTexture the MTex, the gpu texture is
bound in ActivateTexture(unit) and unbound in new function DisableTexture.
As we say, BL_Texture has a function to unbind only one texture which allow us
to remove the slow function DisableAllTextures and other static functions.
By doing this we also remove the environment texture setup unused currently.

The only issue noticed was about texture's bindcode for video texture. Indeed
Multiple MTex can use the same GPUTexture but only one MTex can switch of opengl
texture bindcode. To solve this we reset the bindcode everytime before bind the
texture.
gain about bindcode : the bindcode is saved in a special struct m_savedData used
only to store data to restore at texture deletion.

The BL_Texture initialisation changed too. The constructor now take arguments
previously passed to Init function as the BL_Texture are now created under pointer
and not basic value. Also arguments changed it's now only the MTex, a boll to
know if it's a cubemap and an other bool for mipmapping.

Thanks to Ulysse Martin to his help on this patch.

Commits message list:
- Remove self convertion of textures in blenderplayer.
- Simplify initcubemap code.
- Remove mNeedsDeleted
- Use common BL_Texture::Init(...)
- Remove BL_Texture::mUnit
- Remove BL_TextureObject
- Remove include KX_Camera.h
- Remove pow2 functions
- GetMaxUnits returns MAXTEX
- Use m_variable instead of mVariable in BL_Texture.
- Rename m_texture to m_bindcode
- Convert texture without bind the material.
- Remove m_ok, IsValid and Validate
- Remove environment texture and use GPU_texture_bind
- Remove ActivateFirst and replace DisableAllTextures by DisableTexture
- Cleanup headers include.
- Fix VideoTexture texture bindcode.
- Use GPU_texture_target instead of cached m_type.
- Use pointer for BL_Texture in KX_BlenderMaterial
- Remove IdMode
- Fix invalid Image texture conversion.
- Free texture at material free.