UPBGE: Add python BL_Texture proxy and dynamic uniform texture values.
Previously there was no way to access to material textures. This can be needed
to modificate texture values but this was also impossible.

BL_Texture class is reconverted on a python proxy since the last cleanup that
reduce the code in it. Thess python proxies are accessed from
KX_BlenderMaterial.textures which is a CListWrapper of the texture list.

Each of these contains value needed by the BGE material shaders, which are:
- diffuseIntensity
- diffuseFactor
- specularIntensity
- specularFactor
- parallaxBump
- parallaxStep
- hardness
- mirror
- normal
- emit

Each are read and write float not clamped.

Now we solve the issue about the user interface to textures but these values
showed before should be treated as uniform in the shader (GPU_dynamic_uniform
instead of GPU_uniform). To do this we simple replace it by GPU_dynamic_uniform
with the proper enum constants added in GPU_material.h.

Of course at the game end all original values of the texture are restored
since they are saved in BL_Texture::m_savedData.

Author: Ulysse Martin, Tristan Porteries