fix warning in GetName function + values not set
this patch fixes the bugs I mentionned in my previous patch.
I have some crashes so I have to find if this comes from my patch or not.
+ UI issue (with normal+parallax steps+parallaxbumpscale enabled on
parallax texture this patch fixes the bugs I mentionned in my previous
patch. not related to this branch but related to parallax I guess)