UPBGE: Fix unregister object without blender object.
Previously UnregisterGameObj from SCA_LogicManager was called for every
objects even objects which doesn't have a blender object like the default
camera game object.

To solve it in KX_Scene::NewRemoveObject check if there's a blender
object and to avoid assert raise in debug mode, the UnregisterGameObj
function check if the iterator is not the end of the list before access
to the second item in it.