UPBGE: Fix unbinded primitives after unset material.
This causes that the VBO or VA unset none vertex attribut because
just before in the material deactivation the vertex attribut number
was set to zero. This was only for the last transparency material.
The result was a glitch in the UI or worse, a crash.

To fix it we just invert the call order : first unbind VBO/VA later
unset material.