UPBGE: Implement KX_BlenderMaterial.ambient/specularAlpha and BL_Texture.lodBias
These material values was here but not supported as dynamic uniform (only for specularAlpha)
and not exposed in the python KX_BlenderMaterial API.

The things was more complicated for the texture lod bias. Indeed the lod bias was already
managed but only per images and not per material textures. To switch it requierd to change
the GLSL code to support the lod bias argument in texture2D and textureCube.

Also this option is exposed in the UI under Influences -> Mipmapping, it's not clamped.

As explained the old per image lod bias setting is no longuer used and then removed.
So the commit 8c64707 and ca50b44 are reverted.

Unfornutaly the conversion from old Image lod bias to lod bias in MTex is impossible due
to a lot of dangling pointer in conversion.